MORE BOOTSTRAP ON BODY.  FORM UPGRADE
I have added the bootstrap to my sections in the body of the site.  The
sizing issue has been a problem so I will need to try and find a fix for
the skills and achievements section.  The form at the bottom has been
converted into a more professional looking form although I need to work
a little more on it.  My footer needs looking at next time.